third_party/WebKit/LayoutTests/http/tests/devtools/sources/bezier-swatch-position-expected.txt

Tests that bezier swatches are updated properly in CSS Sources.

Initial swatch positions:
Line 1, Column 25: ease-in-out
Line 4, Column 32: cubic-bezier(0.25, 0.46, 0.45, 0.94)

Running: testEditBezier
Line 1, Column 25: linear
Line 4, Column 32: cubic-bezier(0.25, 0.46, 0.45, 0.94)

Running: testAddBezier
Line 1, Column 25: linear
Line 1, Column 55: cubic-bezier(0, 0.5, 1, 1)
Line 4, Column 32: cubic-bezier(0.25, 0.46, 0.45, 0.94)

Running: testInvalidateBezier
Line 1, Column 25: linear
Line 4, Column 32: cubic-bezier(0.25, 0.46, 0.45, 0.94)
